+++ akshare/__init__.py
@@ -2681,9 +2681,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.12.66 fix: fix stock_board_industry_hist_em interface
 1.12.67 fix: fix index_hist_sw interface
 1.12.68 fix: fix option_finance_board interface
+1.12.69 fix: fix futures_hold_pos_sina interface
 """
 
-__version__ = "1.12.68"
+__version__ = "1.12.69"
 __author__ = "AKFamily"
 
 import sys